Adding postgres for db and testings
