+++ src/blog-env/bin/pip
@@ -1,0 +1,8 @@
+#!/Users/kungfuadmin/projects/tony_fastapi/blog-env/bin/python3.9
+# -*- coding: utf-8 -*-
+import re
+import sys
+from pip._internal.cli.main import main
+if __name__ == '__main__':
+    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
+    sys.exit(main())
