+++ src/blog-env/bin/python3
@@ -1,0 +1,1 @@
+python3.9
