+++ src/blog-env/bin/python3.9
@@ -1,0 +1,1 @@
+/usr/local/opt/python@3.9/bin/python3.9
