+++ src/blog-env/pyvenv.cfg
@@ -1,0 +1,3 @@
+home = /usr/local/opt/python@3.9/bin
+include-system-site-packages = false
+version = 3.9.4
