+++ src/blog/main.py
@@ -1,0 +1,8 @@
+from fastapi import FastAPI
+
+app = FastAPI()
+
+@app.post('/blog')
+def create():
+    return 'creating'
+
